Update patch set 1
Patch Set 1:

Ok, I've looked into the code of jflteatt and I see that the extract-files.sh script for that device creates a jfltei337 directory for proprietary blobs. The problem is: there is a script, in a certain repository, which is meant to create the jf-gsm-common directory? Because, from what I can see, the problem of the missing libsec-ril.so library can be easily recreated also for the jflteatt device. And, by the way, the device-proprietary-files.txt of the jflteatt directory will never be found by the extract-files.sh script in jf-common; it looks for a samsung/jfltei337/device-proprietary-files.txt, it should be samsung/jflteatt/device-proprietary-files.txt.

Patch-set: 1