Update patch set 1
Patch Set 1: Code-Review-1

So, it turned out that this did not work. The image built, but when the resulting jfltexx image was flashed to the jfltetmo the device failed to boot. Instead, it stayed at the initial boot screen (showing the Galaxy S4 logo) and never proceeded. No logs are available from what was going on at the time, so I'm not sure the reason for the failure.

Patch-set: 1