Update patch set 1
Patch Set 1:

> I'll rebuild as eng and see what I find from the logs.

I built an eng build and flashed it. However, as the phone never got past the splash screen I was unable to connect with adb and gather logs. ):

What do you mean by "vendor init"?

Note that if I build a jfltetmo build it is successful and the device does boot.

Patch-set: 1