Update patch set 1
Patch Set 1:

Suspecting that there may be something wrong with the build itself, I attempted to re-create the build from scratch in a clean workspace. I was unable to do so. I guess when I created my initial jfltetmo build I needed to hack a few things to get it to work (mostly with binary blobs), and I'm not able to reproduce this build currently without the hacks. I'll work on getting a clean build using TheMuppets, getting advice on slack. When that is resolved, I'll continue this change.

Patch-set: 1