Update patch set 2
Patch Set 2: Verified-1

I can agree to add back jflte here, since technically that's the correct device name for jfltexx, but it shouldn't be present in other variants. Simply because it will make it possible to flash the wrong firmware, leading to a soft-brick. And that is very very wrong.

And jfltecan is 99.9% identical to jflteatt on the kernel source side, and I337=jflteatt, I337M=jfltecan, so they belong together.

This whole assert/unifying stuff has gotten out of control, and really needs to be cleaned up. If that means forcing users over to new recoveries, so be it. It's better to do it now than letting it get more out of control.

Patch-set: 2
Reviewer: Gerrit User 444 <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=-1